refactor: переместить extended_document в utils
