+++ backend/blogapp/modules/articles/models.py
@@ -6,7 +6,7 @@ from beanie import Link
 from pydantic import BaseModel, model_validator, Field
 
 from ..users.models import UserDocument
-from ...core.database.extended_document import ExtendedDocument
+from ...utils.extended_document import ExtendedDocument
 
 
 class ArticleDocument(ExtendedDocument):
